database created data show in frontend from database

+++ Cars/Models/Car.cs
@@ -11,8 +11,6 @@ namespace Cars.Models
         [ScaffoldColumn(false)]
         public int CarId { get; set; }
 
-       
-
         [Required]
         public int MakeId { get; set; } //   Honda
 
@@ -41,7 +39,7 @@ namespace Cars.Models
         public int InteriorColorId { get; set; }
 
         [Required] 
-        public decimal Price { get; set; }
+        public int Price { get; set; }
 
         
         public string[] CarImage { get; set; }
